requirements.txt
# Member A (AI/NLP)
anthropic>=0.25.0
spacy>=3.9.0
sentence-transformers>=3.0.0
pdfplumber>=0.11.0
python-docx>=1.1.0
numpy>=1.26.0
pydantic>=2.0.0